+++ package.json
@@ -63,6 +63,7 @@
     "@types/lodash": "^4.14.115",
     "@types/mini-css-extract-plugin": "^0.2.0",
     "@types/node": "^8.10.0",
+    "@types/optimize-css-assets-webpack-plugin": "^1.3.3",
     "@types/react": "^16.4.14",
     "@types/react-dom": "^16.0.8",
     "@types/react-redux": "^6.0.9",
@@ -105,6 +106,7 @@
     "moment": "2.22.2",
     "node-sass": "4.9.4",
     "npm-run-all": "4.1.3",
+    "optimize-css-assets-webpack-plugin": "^5.0.1",
     "postcss-flexbugs-fixes": "4.1.0",
     "postcss-loader": "3.0.0",
     "raf": "3.4.0",
